Move menu bar to within the Photo Booth app.
Everything is cursed.
