+++ lib/views/photo_booth_screen/photo_booth.dart
@@ -8,6 +8,7 @@ import 'package:go_router/go_router.dart';
 import 'package:momento_booth/app_localizations.dart';
 import 'package:momento_booth/extensions/go_router_extension.dart';
 import 'package:momento_booth/main.dart';
+import 'package:momento_booth/managers/project_manager.dart';
 import 'package:momento_booth/managers/settings_manager.dart';
 import 'package:momento_booth/utils/logger.dart';
 import 'package:momento_booth/utils/route_observer.dart';
@@ -47,46 +48,54 @@ class PhotoBoothState extends State<PhotoBooth> {
 
   @override
   Widget build(BuildContext context) {
-    return FramerateMonitor(
-      child: LiveViewBackground(
-        router: _router,
-        child: _HotkeyResponder(
-          router: _router,
-          child: ActivityMonitor(
-            router: _router,
-            child: MomentoBoothTheme(
-              data: MomentoBoothThemeData.defaults(),
-              child: SetScrollConfiguration(
-                child: Observer(
-                  builder: (context) => FluentApp.router(
-                    debugShowCheckedModeBanner: false,
-                    scrollBehavior: ScrollConfiguration.of(context),
-                    color: getIt<SettingsManager>().settings.ui.primaryColor,
-                    theme: FluentThemeData(
-                      accentColor: AccentColor.swatch(
-                        {'normal': getIt<SettingsManager>().settings.ui.primaryColor},
+    return Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          _menuBar(context, _router),
+          Expanded(
+            child: FramerateMonitor(
+              child: LiveViewBackground(
+                router: _router,
+                child: _HotkeyResponder(
+                  router: _router,
+                  child: ActivityMonitor(
+                    router: _router,
+                    child: MomentoBoothTheme(
+                      data: MomentoBoothThemeData.defaults(),
+                      child: SetScrollConfiguration(
+                        child: Observer(
+                          builder: (context) => FluentApp.router(
+                            debugShowCheckedModeBanner: false,
+                            scrollBehavior: ScrollConfiguration.of(context),
+                            color: getIt<SettingsManager>().settings.ui.primaryColor,
+                            theme: FluentThemeData(
+                              accentColor: AccentColor.swatch(
+                                {'normal': getIt<SettingsManager>().settings.ui.primaryColor},
+                              ),
+                            ),
+                            routerConfig: _router,
+                            localizationsDelegates: const [
+                              AppLocalizations.delegate,
+                              GlobalMaterialLocalizations.delegate,
+                              GlobalWidgetsLocalizations.delegate,
+                              GlobalCupertinoLocalizations.delegate,
+                              FluentLocalizations.delegate,
+                            ],
+                            supportedLocales: const [
+                              Locale('en'), // English
+                              Locale('nl'), // Dutch
+                            ],
+                            locale: getIt<SettingsManager>().settings.ui.language.toLocale(),
+                          ),
+                        ),
                       ),
                     ),
-                    routerConfig: _router,
-                    localizationsDelegates: const [
-                      AppLocalizations.delegate,
-                      GlobalMaterialLocalizations.delegate,
-                      GlobalWidgetsLocalizations.delegate,
-                      GlobalCupertinoLocalizations.delegate,
-                      FluentLocalizations.delegate,
-                    ],
-                    supportedLocales: const [
-                      Locale('en'), // English
-                      Locale('nl'), // Dutch
-                    ],
-                    locale: getIt<SettingsManager>().settings.ui.language.toLocale(),
                   ),
                 ),
               ),
             ),
           ),
-        ),
-      ),
+        ]
     );
   }
 
@@ -95,4 +104,48 @@ class PhotoBoothState extends State<PhotoBooth> {
     _router.dispose();
     super.dispose();
   }
+}
+
+Widget _menuBar(BuildContext context, GoRouter router) {
+  return ColoredBox(
+    color: Color(0xFFFFFFFF),
+    child: MenuBar(
+      items: [
+        MenuBarItem(title: 'File', items: [
+          MenuFlyoutSubItem(
+            text: const Text('Recent projects'),
+            items: (context) {
+              return [
+                MenuFlyoutItem(
+                  text: const Text('Plain Text Documents'),
+                  onPressed: () {},
+                ),
+                MenuFlyoutItem(
+                  text: const Text('Rich Text Documents'),
+                  onPressed: () {},
+                ),
+                MenuFlyoutItem(
+                  text: const Text('Other Formats'),
+                  onPressed: () {},
+                ),
+              ];
+            },
+          ),
+          MenuFlyoutItem(text: const Text('Open'), onPressed: getIt<ProjectManager>().browseOpen),
+          MenuFlyoutItem(text: const Text('Settings'), onPressed: () { router.push("/settings"); }),
+          const MenuFlyoutSeparator(),
+          MenuFlyoutItem(text: const Text('Exit'), onPressed: () {}),
+        ]),
+        MenuBarItem(title: 'Edit', items: [
+          MenuFlyoutItem(text: const Text('Undo'), onPressed: () {}),
+          MenuFlyoutItem(text: const Text('Cut'), onPressed: () {}),
+          MenuFlyoutItem(text: const Text('Copy'), onPressed: () {}),
+          MenuFlyoutItem(text: const Text('Paste'), onPressed: () {}),
+        ]),
+        MenuBarItem(title: 'Help', items: [
+          MenuFlyoutItem(text: const Text('About'), onPressed: () {}),
+        ]),
+      ],
+    ),
+  );
 }
